Use join delimiter for commands and config files
Specify a join delimiter only once instead of manually adding spaces or newlines on the ends of joined strings. This is less error prone for building commands and configuration files as one does not risk forgetting one. For configuration files it also makes the content more readable.

Not applicable for user data content which needs to mix newlines to separate commands and spaces to insert computed values within commands.